requirements.txt
logging
sqlite3
os
requests
json
unidecode
pygsheets
mysql-connector-python